Add reconnection and resend mechanism for TCP
When server closes the socket, it takes two write to the socket for the
client to realize that the connection no longer exists or was broken.
The first write (after the server closes the connection) returns success
and the second write returns EPIPE error.

Since it's not possible to immediately detect that the server closed the
connection, this commit includes two mechanisms that would avoid data
lost when it counts.

* First mechanism tries to reconnect to the server after EPIPE error.
  If the server is still down after the attemp to reconnect, the write
  handle returns "Connection refused" and and the message is cleared.
* Second mechanism keeps track of the previous sent message so that when
  EPIPE happens, it can resend the first message that was not sent
  successfully. This mechanism has a timeout perid of 10 seconds so that
  When the interval between the first and second messages are more than
  10 seconds, only the second message is sent.