Attempt to fix "unload" not sent on Linux on exit
It seems like the problem has to do when user enters CTL-D to exit the
program. Since CTL-D does not send a signal but enters EOF in stdin,
something goes on that make the program end adruptly without calling the
destructor for AssetManagerClient.

The patch calls fclose(stdin) before exiting the main function and this
seems to fix the above problem on Linux (CentOS 5.7)