Remove unused conditional variable
The problem of "unload" not getting sent on Linux depending on quiting
by entering CTL-C or CTL-D is still unsolved. However, a boolean
variable for the main while loop was not being used (because the loop
is broken when eof() is detected on stdin), this commit removes the
unecessary variable.

To fix the "unload" not getting sent problem, it's probably best to
change the interface for AssetManagerClient and explicitly send Unload()
and explicitly block to send all the remaining messages rather than
including this time consuming process in the destructor.