Move blocking part in destructor to a new function
Destructor was designed to block to send all remaining messages in the
queue, but since destructor should do minimal amount of work, this was
not the right way to go. All the blocking portion was moved to a new
function BlockUntilQueuesAreEmpty, which blocks until all the messages
in the TCP and UDP queues are processed.

Moreover, option for sending the unload message was removed. This should
be called explicitly by the user to have the right behavior.

Tests program (main_test) and CLI program (am_client) are also updated
to reflect the new API change.