Fix BlockUntilQueueIsEmpty bug
BlockUntilQueueIsEmpty did not block properly when a new message was
passed to "Send()" but the client thread did not reach "DoSend()" before
BlockUntilQueueIsEmpty was called. Additional boolean was added to make
sure there is a message that needs to be handled and
BlockUntilQueueIsEmpty will block until that message gets processed.

Updated the test program to better demonstrate the use of the API.